+++ liana/plotting/__init__.py
@@ -1,1 +1,1 @@
-from ._dotplot import dotplot
+from ._dotplot import dotplot, dotplot_by_sample