+++ liana/plotting/_dotplot.py
@@ -68,41 +68,7 @@ def dotplot(adata: anndata.AnnData = None,
     A `plotnine.ggplot` instance
 
     """
-    if (liana_res is not None) & (adata is not None):
-        raise AttributeError('Ambiguous! One of `liana_res` or `adata` should be provided.')
-    if adata is not None:
-        assert 'liana_res' in adata.uns_keys()
-        liana_res = adata.uns['liana_res'].copy()
-    if liana_res is not None:
-        liana_res = liana_res.copy()
-    if (liana_res is None) & (adata is None):
-        raise ValueError('`liana_res` or `adata` must be provided!')
-    if (colour is None) | (size is None):
-        raise ValueError('`colour` and `size` must be provided!')
-    if colour not in liana_res.columns:
-        raise ValueError('`colour` column was not found in `liana_res`!')
-    if size not in liana_res.columns:
-        raise ValueError('`size` column was not found in `liana_res`!')
-
-    liana_res['interaction'] = liana_res.ligand_complex + ' -> ' + liana_res.receptor_complex
-
-    # subset to only cell labels of interest
-    if source_labels is not None:
-        source_msk = np.isin(liana_res.source, source_labels)
-        liana_res = liana_res[source_msk]
-        possible_sources = np.unique(liana_res['source'])
-        covered = np.isin(source_labels, possible_sources)
-        if not all(covered):
-            not_covered = np.array(source_labels)[~covered]
-            raise ValueError(f"{not_covered} not found in `liana_res['source']`!")
-    if target_labels is not None:
-        target_msk = np.isin(liana_res.target, target_labels)
-        liana_res = liana_res[target_msk]
-        possible_targets = np.unique(liana_res['target'])
-        covered = np.isin(target_labels, possible_targets)
-        if not all(covered):
-            not_covered = np.array(target_labels)[~covered]
-            raise ValueError(f"{not_covered} not found in `liana_res['target']`!")
+    liana_res = _prep_liana_res(adata=adata, liana_res=liana_res, source_labels=source_labels, target_labels=target_labels)
 
     if filterby is not None:
         msk = liana_res[filterby].apply(filter_lambda)
@@ -162,6 +128,73 @@ def dotplot(adata: anndata.AnnData = None,
 
     p.draw()
 
+    
+def dotplot_by_sample(adata=None,
+                      liana_res=None,
+                      colour=None,
+                      size=None,
+                      source_labels=None,
+                      target_labels=None,
+                      ligand_complex=None, 
+                      receptor_complex=None,
+                      sample_key='sample',
+                      size_range: tuple = (2, 9),
+                      figure_size: tuple = (8, 6),
+                      return_fig=True):
+    
+    liana_res = _prep_liana_res(adata=adata, liana_res=liana_res, source_labels=source_labels, target_labels=target_labels)
+    
+    if ligand_complex is not None:
+        liana_res = liana_res[np.isin(liana_res['ligand_complex'], ligand_complex)]
+    if receptor_complex is not None:
+        liana_res = liana_res[np.isin(liana_res['receptor_complex'], receptor_complex)]
+
+    p = (ggplot(liana_res, aes(x='target', y='source', colour=colour, size=size))
+            + geom_point()
+            + facet_grid(f'interaction~{sample_key}')
+            + scale_size_continuous(range=size_range)
+            + labs(color=str.capitalize(colour),
+                   size=str.capitalize(size),
+                   y="Source",
+                   x="Target",
+                   title=sample_key.capitalize())
+            + theme_bw()
+            + theme(legend_text=element_text(size=14),
+                    strip_background=element_rect(fill="white"),
+                    strip_text=element_text(size=13, colour="black"),
+                    axis_text_y=element_text(size=10, colour="black"),
+                    axis_title_y=element_text(colour="#808080", face="bold", size=14),
+                    axis_text_x=element_text(size=11, face="bold", angle=90),
+                    axis_title_x=element_text(colour="#808080", face="bold", size=14),
+                    figure_size=figure_size,
+                    plot_title=element_text(vjust=0, hjust=0.5, face="bold", size=15),
+                    )
+            )
+    if return_fig:
+        return p
+    
+    p.draw()
+
+
+def _prep_liana_res(adata=None, liana_res=None, source_labels=None, target_labels=None):
+    if (liana_res is None) & (adata is None):
+        raise AttributeError('Ambiguous! One of `liana_res` or `adata` should be provided.')
+    if adata is not None:
+        assert 'liana_res' in adata.uns_keys()
+        liana_res = adata.uns['liana_res'].copy()
+    if liana_res is not None:
+        liana_res = liana_res.copy()
+    if (liana_res is None) & (adata is None):
+        raise ValueError('`liana_res` or `adata` must be provided!')
+
+    # subset to only cell labels of interest
+    liana_res = _filter_labels(liana_res, labels=source_labels, label_type='source')
+    liana_res = _filter_labels(liana_res, labels=target_labels, label_type='target')
+    
+    liana_res['interaction'] = liana_res['ligand_complex'] + ' -> ' + liana_res['receptor_complex']
+
+    return liana_res
+
 
 def _aggregate_scores(res, what, how, entities):
     return res.groupby(entities).agg(score=(what, how)).reset_index()
@@ -169,3 +202,18 @@ def _aggregate_scores(res, what, how, entities):
 
 def _inverse_scores(score):
     return -np.log10(score + np.finfo(float).eps)
+
+
+def _filter_labels(liana_res, labels, label_type):
+    if labels is not None:
+        if labels is str:
+            labels = [labels]
+        covered = np.isin(labels, liana_res[label_type])
+        if not covered.all():
+            not_covered = np.array(labels)[~covered]
+            raise ValueError(f"{not_covered} not found in `liana_res['{label_type}']`!")
+        msk = np.isin(liana_res[label_type], labels)
+        liana_res = liana_res[msk]
+        
+    return liana_res
+